implement range functions (crudely), and fini function

+++ mt19937/include/mt19937/random.h
@@ -10,28 +10,29 @@
 
 /* STATE MODULE */
 
+#include <uni/types/int.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 struct prng;
 
-struct mt_prng* mt_prng_init( void );
-void mt_prng_fini( struct mt_prng* );
+struct mt_prng *mt_prng_init(void);
+void mt_prng_fini(struct mt_prng *);
 
-int mt_range_s32( struct mt_prng*, int, int );
-unsigned mt_range_u32( struct mt_prng*, unsigned, unsigned );
-#if !defined( CFG_GBA )
-long long mt_range_s64( struct mt_prng*, long long, long long );
-unsigned long long mt_range_u64(
-	struct mt_prng*, unsigned long long, unsigned long long );
+s32 mt_range_s32(struct mt_prng *, s32, s32);
+u32 mt_range_u32(struct mt_prng *, u32, u32);
+#if !defined(CFG_GBA)
+s64 mt_range_s64(struct mt_prng *, s64, s64);
+u64 mt_range_u64(struct mt_prng *, u64, u64);
 #endif
 
-int mt_random_s32( struct mt_prng* );
-unsigned mt_random_u32( struct mt_prng* );
-#if !defined( CFG_GBA )
-long long mt_random_s64( struct mt_prng* );
-unsigned long long mt_random_u64( struct mt_prng* );
+s32 mt_random_s32(struct mt_prng *);
+u32 mt_random_u32(struct mt_prng *);
+#if !defined(CFG_GBA)
+s64 mt_random_s64(struct mt_prng *);
+u64 mt_random_u64(struct mt_prng *);
 #endif
 
 #ifdef __cplusplus
